requirements.txt
keras
numpy
selenium
Django
django-rest-auth
djangorestframework
django-session-timeout